setActiveStatus logic big refactor: split to small pieces to make it reusable
